requirements.txt
pandas==2.3.1
typing_extensions==4.14.1
python-dotenv==1.2.1